perf(rust, python): optimize string kernels, (elide redundant allocs)
