+++ polars/polars-ops/src/chunked_array/strings/mod.rs
@@ -1,5 +1,9 @@
+#[cfg(feature = "strings")]
+mod case;
 #[cfg(feature = "extract_jsonpath")]
 mod json_path;
+#[cfg(feature = "string_justify")]
+mod justify;
 #[cfg(feature = "strings")]
 mod namespace;
 #[cfg(feature = "strings")]
